+++ src/common/services/pmsearch/PmSearchApiService.ts
@@ -1,8 +1,9 @@
 import { BackendSrv, BackendSrvRequest, FetchResponse } from '@grafana/runtime';
 import { defaults, has } from 'lodash';
-import { DefaultRequestOptions, getRequestOptions, timeout, TimeoutError } from '../../../common/utils';
+import { firstValueFrom } from 'rxjs';
 import { SearchEntityUtil } from '../../../components/search/utils/SearchEntityUtil';
 import { NetworkError } from '../../types/errors';
+import { DefaultRequestOptions, getRequestOptions, timeout, TimeoutError } from '../../utils';
 import {
     AutocompleteQueryParams,
     AutocompleteResponse,
@@ -23,7 +24,7 @@ export class PmSearchApiService {
     async request<T>(options: BackendSrvRequest): Promise<FetchResponse<T>> {
         options = defaults({}, options, this.defaultRequestOptions);
         try {
-            return await timeout(this.backendSrv.fetch<T>(options).toPromise(), this.apiConfig.timeoutMs);
+            return await timeout(firstValueFrom(this.backendSrv.fetch<T>(options)), this.apiConfig.timeoutMs);
         } catch (error) {
             if (error instanceof TimeoutError) {
                 throw new TimeoutError(`Timeout while connecting to '${options.url}'`, error);
@@ -45,7 +46,7 @@ export class PmSearchApiService {
         try {
             const response = await this.request<AutocompleteResponse>(request);
             return response.data;
-        } catch (error) {
+        } catch (error: any) {
             if (has(error, 'data.success') && !error.data.success && Object.keys(error.data).length === 1) {
                 throw new SearchNotAvailableError();
             } else {
@@ -77,7 +78,7 @@ export class PmSearchApiService {
                     offset: params.offset ?? 0,
                 };
             }
-        } catch (error) {
+        } catch (error: any) {
             if (has(error, 'data.success') && !error.data.success && Object.keys(error.data).length === 1) {
                 throw new SearchNotAvailableError();
             } else {
@@ -115,7 +116,7 @@ export class PmSearchApiService {
                     offset: params.offset ?? 0,
                 };
             }
-        } catch (error) {
+        } catch (error: any) {
             if (has(error, 'data.success') && !error.data.success && Object.keys(error.data).length === 1) {
                 throw new SearchNotAvailableError();
             } else {